+++ src/main/java/com/example/waystoneinjector/WaystoneInjectorMod.java
@@ -3,8 +3,6 @@ package com.example.waystoneinjector;
 import com.example.waystoneinjector.config.WaystoneConfig;
 import net.minecraftforge.fml.common.Mod;
 import net.minecraftforge.fml.loading.FMLEnvironment;
-import net.minecraftforge.eventbus.api.IEventBus;
-import net.minecraftforge.fml.javafmlmod.FMLJavaModLoadingContext;
 
 @Mod(WaystoneInjectorMod.MODID)
 public class WaystoneInjectorMod {
@@ -15,6 +13,9 @@ public class WaystoneInjectorMod {
         if (FMLEnvironment.dist.isClient()) {
             // Register config
             WaystoneConfig.register();
+            
+            // Register Feverdream packet listener
+            com.example.waystoneinjector.network.FeverdreamNetworking.register();
             
             // Register client-only event handlers
             net.minecraftforge.common.MinecraftForge.EVENT_BUS.register(new com.example.waystoneinjector.client.ClientEvents());